extract wraputils.lua from splash.lua
